requirements.txt
Pillow>=9.0
pywebview>=4.0
